Move quotes array to a separate file
