+++ src/Pages/Quotes/Quotes.jsx
@@ -1,35 +1,7 @@
 import './quotes.scss';
+import quotes from '../../Data/quotes';
 
 const Quotes = () => {
-  const quotes = [
-    'Plato-Geometry existed before creation.',
-    'Plato-God forever geometrizes.',
-    'Proclus-Wherever there is number, there is beauty.',
-    'Galileo-The book of nature is written in the language of mathematics.',
-    'Carl Friedrich Gauss-Mathematics is the Queen of the Sciences.',
-    'Consider Sterry-But of all the sciences cultivated by mankind, none are more useful than mathematics, to call forth a spirit of enterprise and enquiry.',
-    'Henri Poincare-The mathematician does not study pure mathematics because it is useful, he studies it because he delights in it and he delights in it because it is beautiful.',
-    'Albert Einstein-Pure mathematics is, in its way, the poetry of logical ideas.',
-    'Albert Einstein-How can it be that mathematics, a product of human thought in dependent of experience, is so admirably adapted to the objects of reality?',
-    'Martin Gardner-All mathematicians share ... a sense of amazement over the infinite depth and the mysterious beauty and usefulness of mathematics.',
-    'Janos Bolyai-Out of Nothing, I have created a strange new universe.',
-    'Bertrand Russell-Mathematics, rightly viewed, possesses not only truth, but supreme beauty.',
-    'Bertrand Russell-Mathematics takes us still further from what is human, into the region of absolute necessity, to which not only the actual world, but every possible world, must conform.',
-    'Bertrand Russell-Math is the subject in which we never know what we are talking about, nor whether what we are saying is true.',
-    'Roger Bacon-Mathematics is the door and key to the sciences.',
-    'Galileo-Mathematics is the language with which God wrote the universe.',
-    'Martin Gardner-Mathematics is not only real, but it is the only reality.',
-    'Descartes-Mathematics is a more powerful instrument of knowledge than any other that has been bequeathed to us by human agency.',
-    'Shakuntala Devi-Without mathematics, there’s nothing you can do. Everything around you is mathematics. Everything around you is numbers.',
-    'Paul Halmos-The only way to learn mathematics is to do mathematics.',
-    'James Joseph Sylvester-Mathematics is the music of reason',
-    'Dean Schlicter-Go down deep enough into anything and you will find mathematics',
-    'Roger Bacon-Mathematics is the gate and key to science',
-    'Snoop Dogg-If you stop at general math, then you will only make general money',
-    'Stefan Banach-Mathematics is the most beautiful and most powerful creation of the human spirit',
-    'Anonymous-Math is the source code of reality',
-  ];
-
   const takeRandomQuote = (quotes, start = 0) => {
     const index = Math.floor(Math.random() * (quotes.length - start + 1) + start);
     return quotes[index];
